third_party/WebKit/LayoutTests/fast/speech/speech-input-scripting-expected.txt

CONSOLE WARNING: The 'x-webkit-speech' input field attribute is deprecated. Please use the JavaScript API instead.
Tests for enabling and disabling speech input via script.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById("speechInput").value is "Pictures of the moon"
PASS document.getElementById("speechInput").value is "Pictures of the moon"
PASS document.getElementById("speechInput").value is "Pictures of the moon"
PASS document.getElementById("speechInput").value is "Pictures of the moon"
PASS successfullyParsed is true

TEST COMPLETE
   
